refactor: move email-parser
